+++ packages/expo-ui/CHANGELOG.md
@@ -18,6 +18,7 @@
 ### 💡 Others
 
 - [apple] Migrate remaining `expo-module.config.json` to unified platform syntax. ([#34445](https://github.com/expo/expo/pull/34445) by [@reichhartd](https://github.com/reichhartd))
+- Rename the events for the `Switch` component.
 
 ## 0.0.1 — 2025-01-21